Set JAX_DEFAULT_DTYPE_BITS env variable to 32 before any imports of JAX
To maintain the default precision as 32-bit and not switch to 64-bit,
we set the JAX_DEFAULT_DTYPE_BITS flag prior to any imports of JAX.
This is necessary as the jax_enable_x64 flag is later set to True inside
the Lanczos algorithm.
See: https://github.com/google/jax/issues/8178 for more information.