Add the option to call the ESE procedure from the main training loop (JAX)
Add the option to call the ESE procedure from the main training loop,
rather than from within the optimizer’s 'update' method (relevant only for
FOSI JAX).

To simplify usage, the ESE procedure can be called internally in the
update function. However, the jax.lax.cond increases the latency, even
when the condition is not satisfied, since the 'false_func' must receive
and return the same types as the 'true_func', which include the state that
is large. (This is also the case, behind the scenes, when using
lax.while_loop, lax.fori_loop, and lax.switch.)

Alternatively, the user could set b_call_ese_internally to False and call
state.update_ese externally. To obtain the best performance, make sure
this is not called from within a jitted function, but rather directly
from the main training loop, and to not involve any device variables - only
cpu variables – when checking the condition. An example could be found in
transfer_learning_cifar10.py.